requirements.txt
torch>=1.12.0
torch-geometric>=2.0.0
biopython>=1.79
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0
tqdm>=4.62.0
matplotlib>=3.5.0

